requirements.txt
asgiref==3.8.1
Django==5.1.7
django-axes==7.1.0
django-resized==1.0.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
